requirements.txt
python-telegram-bot>=20.0
Flask>=2.0
requests>=2.25
python-dotenv>=0.19
gTTS>=2.2
